Tests/LibWeb/Text/expected/DOMTokenList-supports.txt

a.relList.supports("alternate"): false
a.relList.supports("stylesheet"): false
a.relList.supports("preload"): false
a.relList.supports("dns-prefetch"): false
a.relList.supports("preconnect"): false
a.relList.supports("icon"): false
a.relList.supports("STYLESHEET"): false
a.relList.supports("never-supported"): false
area.relList.supports("alternate"): false
area.relList.supports("stylesheet"): false
area.relList.supports("preload"): false
area.relList.supports("dns-prefetch"): false
area.relList.supports("preconnect"): false
area.relList.supports("icon"): false
area.relList.supports("STYLESHEET"): false
area.relList.supports("never-supported"): false
form.relList.supports("alternate"): false
form.relList.supports("stylesheet"): false
form.relList.supports("preload"): false
form.relList.supports("dns-prefetch"): false
form.relList.supports("preconnect"): false
form.relList.supports("icon"): false
form.relList.supports("STYLESHEET"): false
form.relList.supports("never-supported"): false
link.relList.supports("alternate"): true
link.relList.supports("stylesheet"): true
link.relList.supports("preload"): true
link.relList.supports("dns-prefetch"): true
link.relList.supports("preconnect"): true
link.relList.supports("icon"): true
link.relList.supports("STYLESHEET"): true
link.relList.supports("never-supported"): false
html.classList.supports("foo") throws: TypeError
